% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance_change.r
\name{calculateVariance}
\alias{calculateVariance}
\title{Calculate change in variance}
\usage{
calculateVariance(
  avt_dat,
  top_genera,
  out_dir = "avt_out",
  top_genera_num2 = 200,
  top_eigen = 10,
  seed = 123
)
}
\arguments{
\item{avt_dat}{List. Output from avt_data().}

\item{top_genera}{List. Output from topGenera().}

\item{out_dir}{String. File path to output directory.}

\item{top_genera_num2}{Numeric. Number of genus to consider. Default set to 200.}

\item{top_eigen}{Numeric. Number of top eigen vectors to consider. Default set to 10.}

\item{seed}{Numeric.}
}
\description{
this function takes the output from organised data and top genus list
and calculate the variance in the top genus
and returns the sum of the top eigen values
}
