% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_genus.r
\name{cleanGenus}
\alias{cleanGenus}
\title{Remove poorly defined Genus from the taxonomy table}
\usage{
cleanGenus(x, y = vector())
}
\arguments{
\item{x}{Taxonomy dataframe.}

\item{y}{A vector with of Genus that should be ignored.}
}
\description{
Clean up Genus to remove poorly defined genus
}
