% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_variance_change.r
\name{orderedVarianceChange}
\alias{orderedVarianceChange}
\title{Summarise the change in variance}
\usage{
orderedVarianceChange(avt_var, out_dir = "avt_out")
}
\arguments{
\item{avt_var}{List. Output from calculateVariance().}

\item{out_dir}{String. File path to output directory.}
}
\description{
Summarise the change in variance
}
