% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eige_sum.r
\name{plotEigeSum}
\alias{plotEigeSum}
\title{Plot the change in variance}
\usage{
plotEigeSum(avt_var, out_dir, height = 6, width = 10)
}
\arguments{
\item{avt_var}{Numeric vector.}

\item{out_dir}{Numeric vector.}

\item{height}{Numeric vector.}

\item{width}{Numeric vector.}
}
\description{
Plot the change in variance
}
