% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top_genus.r
\name{topByCentrality}
\alias{topByCentrality}
\title{Find the top Genus by centrality}
\usage{
topByCentrality(
  mean_genus_count,
  taxa,
  out_dir,
  current_grouping_name,
  top_number,
  centrality_p,
  centrality_r_upper,
  centrality_r_lower
)
}
\arguments{
\item{mean_genus_count}{Numeric vector.}

\item{taxa}{Numeric vector.}

\item{out_dir}{Numeric vector.}

\item{current_grouping_name}{Numeric vector.}

\item{top_number}{Numeric vector.}

\item{centrality_p}{Numeric vector.}

\item{centrality_r_upper}{Numeric vector.}

\item{centrality_r_lower}{Numeric vector.}
}
\description{
Find the top Genus by centrality
}
