% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top_genus.r
\name{topByPrevalence}
\alias{topByPrevalence}
\title{Find the top Genus by prevalence}
\usage{
topByPrevalence(mean_genus_count, top_number)
}
\arguments{
\item{mean_genus_count}{Numeric vector.}

\item{top_number}{Numeric vector.}
}
\description{
Find the top Genus by prevalence
}
