% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top_genera.r
\name{topGenera}
\alias{topGenera}
\title{Determine the top Genus (by abundance, prevalence or centrality)}
\usage{
topGenera(
  avt_dat,
  out_dir = "avt_out",
  top_number = 50,
  centrality_p = 0.05,
  centrality_r_upper = 0.7,
  centrality_r_lower = 0
)
}
\arguments{
\item{avt_dat}{List. Output from avt_data().}

\item{out_dir}{String. File path to output directory.}

\item{top_number}{Numeric. Top number of taxa to include. Default = 50.}

\item{centrality_p}{Numeric. The probablity threahold for Pearson correlation. Default value = 0.05}

\item{centrality_r_upper}{Numeric. The upper level of Pearson correlation (r) to consider for centrality calculation. Include r above this value. Default = 0.7.}

\item{centrality_r_lower}{Numeric. The lower level of Pearson correlation (r) to consider for centrality calculation. Include r below this value. Default = 0}
}
\description{
Determine the top Genus (by abundance, prevalence or centrality)
}
