% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top_genus.r
\name{topGenus}
\alias{topGenus}
\title{Determine the top Genus (by abundance, prevalence or centrality)}
\usage{
topGenus(
  avt_dat,
  out_dir,
  top_number = 50,
  centrality_p = 0.05,
  centrality_r_upper = 0.7,
  centrality_r_lower = 0
)
}
\arguments{
\item{avt_dat}{Numeric vector.}

\item{out_dir}{Numeric vector.}

\item{top_number}{Numeric vector.}

\item{centrality_p}{Numeric vector.}

\item{centrality_r_upper}{Numeric vector.}

\item{centrality_r_lower}{Numeric vector.}
}
\description{
Determine the top Genus (by abundance, prevalence or centrality)
}
