% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_organisation.r
\name{avt_data}
\alias{avt_data}
\title{Organise input data into AVT format}
\usage{
avt_data(
  data_dir,
  grouping_column,
  out_dir,
  count_filename = "OTU.txt",
  taxa_filename = "taxonomy.tsv",
  meta_filename = "MetaDB.txt",
  tree_filename = "tree.nwk",
  f_taxaconfidence = 0.9,
  f_grouping_inclusion = vector(),
  min_req_sample_number = 8
)
}
\arguments{
\item{data_dir}{Numeric vector.}

\item{grouping_column}{Numeric vector.}

\item{out_dir}{Numeric vector.}

\item{count_filename}{Numeric vector.}

\item{taxa_filename}{Numeric vector.}

\item{meta_filename}{Numeric vector.}

\item{tree_filename}{Numeric vector.}

\item{f_taxaconfidence}{Numeric vector.}

\item{f_grouping_inclusion}{Numeric vector.}

\item{min_req_sample_number}{Numeric vector.}
}
\description{
This tool tracks microbiome variance and identifies bacterial genera responsible for significant variance shifts using 16S rRNA gene metabarcoding for longitudinal data.
}
