% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance_change.r
\name{calculateVariance}
\alias{calculateVariance}
\title{Calculate change in variance}
\usage{
calculateVariance(
  avt_dat,
  top_genus,
  out_dir,
  top_genus_num2 = 200,
  top_eigen = 5
)
}
\arguments{
\item{avt_dat}{Numeric vector.}

\item{top_genus}{Numeric vector.}

\item{out_dir}{Numeric vector.}

\item{top_genus_num2}{Numeric vector.}

\item{top_eigen}{Numeric vector.}
}
\description{
this function takes the output from organised data and top genus list
and calculate the variance in the top genus
and returns the sum of the top eigen values
}
