% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eige_sum.r
\name{plotEigeSum}
\alias{plotEigeSum}
\title{Plot the change in variance}
\usage{
plotEigeSum(avt_var, out_dir = "avt_out", height = 6, width = 10)
}
\arguments{
\item{avt_var}{List. Output from calculateVariance().}

\item{out_dir}{String. File path to output directory.}

\item{height}{Numeric. Height of the output PDF plot.}

\item{width}{Numeric. Width of the output PDF plot.}
}
\description{
Plot the change in variance
}
